requirements.txt
arrow==0.16.0
django==3.1.2
django-environ==0.4.5
feedparser==6.0.1
httpx==0.15.4
Jinja2==2.11.2
Mastodon.py==1.5.1
praw==7.1.0
pypandoc
python-slugify
rich==8.0.0
